Create rules folder when downloading
In the download command, there is no check to create the ```/rules``` folder if it is not already present. However, the code has been written to always write the file to this folder.

So when a user does specify the folder name to use, an error is generated saying ```No such file or directory```. The fix I've made is to correct this bug.